+++ sdk-api-infinity/src/main/kotlin/com/pexip/sdk/api/infinity/internal/RealCallStep.kt
@@ -27,18 +27,18 @@ import com.pexip.sdk.api.infinity.UpdateRequest
 import com.pexip.sdk.api.infinity.UpdateResponse
 import kotlinx.serialization.SerializationException
 import kotlinx.serialization.json.Json
+import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.Response
 import okhttp3.internal.EMPTY_REQUEST
-import java.net.URL
 import java.util.UUID
 import java.util.concurrent.TimeUnit
 
 internal class RealCallStep(
     private val client: OkHttpClient,
     private val json: Json,
-    private val node: URL,
+    private val node: HttpUrl,
     private val conferenceAlias: String,
     private val participantId: UUID,
     private val callId: UUID,