+++ sdk-api-infinity/src/main/kotlin/com/pexip/sdk/api/infinity/internal/RealConferenceStep.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Pexip AS
+ * Copyright 2022-2023 Pexip AS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -32,17 +32,17 @@ import com.pexip.sdk.api.infinity.SsoRedirectException
 import com.pexip.sdk.api.infinity.Token
 import kotlinx.serialization.SerializationException
 import kotlinx.serialization.json.Json
+import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.Response
 import okhttp3.internal.EMPTY_REQUEST
-import java.net.URL
 import java.util.UUID
 
 internal class RealConferenceStep(
     private val client: OkHttpClient,
     private val json: Json,
-    private val node: URL,
+    private val node: HttpUrl,
     private val conferenceAlias: String,
 ) : InfinityService.ConferenceStep {