+++ sdk-api-infinity/src/main/kotlin/com/pexip/sdk/api/infinity/internal/RealParticipantStep.kt
@@ -27,17 +27,17 @@ import com.pexip.sdk.api.infinity.NoSuchNodeException
 import com.pexip.sdk.api.infinity.Token
 import kotlinx.serialization.SerializationException
 import kotlinx.serialization.json.Json
+import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.Response
 import okhttp3.internal.EMPTY_REQUEST
-import java.net.URL
 import java.util.UUID
 
 internal class RealParticipantStep(
     private val client: OkHttpClient,
     private val json: Json,
-    private val node: URL,
+    private val node: HttpUrl,
     private val conferenceAlias: String,
     private val participantId: UUID,
 ) : InfinityService.ParticipantStep {
@@ -216,15 +216,14 @@ internal class RealParticipantStep(
     override fun message(request: MessageRequest, token: Token): Call<Boolean> =
         message(request, token.token)
 
-    override fun call(callId: UUID): InfinityService.CallStep =
-        RealCallStep(
-            client = client,
-            json = json,
-            node = node,
-            conferenceAlias = conferenceAlias,
-            participantId = participantId,
-            callId = callId,
-        )
+    override fun call(callId: UUID): InfinityService.CallStep = RealCallStep(
+        client = client,
+        json = json,
+        node = node,
+        conferenceAlias = conferenceAlias,
+        participantId = participantId,
+        callId = callId,
+    )
 
     private fun parseCalls(response: Response) = when (response.code) {
         200 -> json.decodeFromResponseBody(CallsResponseSerializer, response.body!!)